combined e2e tests and removed some noise from empty tests
